+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# 1.0.0
+- Enable negative `lengthAngle` to configure clockwise and counterclockwise charts
+
+- Make counterclockwise charts by default
+
+# 0.0.2
+- Incorporate `svg-partial-circle` lib
+
 # 0.0.1
 - Fix bad imports